feat(upgrade): api/upgrade entry — union recompose, version-sync restart retained, shim retired
Move composite CheckResult assembly from update-checker.ts shim into
daemon/api/upgrade.ts (the new union recompose owner). Routes renamed
/api/update/* → /api/upgrade/*; UI fetch paths updated to match.

myco apply path: initiateAdopt on pre-staged binary (422 when none staged).
Operator CLI path: spawnUpdateScript (npm) unchanged.
Version-sync self-restart (spawnRestartScript + restarting/reason) retained.

Carry-forwards resolved: runtimeCommand dead param dropped, PackageCheckResult
consolidated to upgrade/checker.ts canonical, latest_stable ?? currentVersion
parity applied. UpdatePackageId import restored in update-checker.ts (tsc gate).

Tests: upgrade.test.ts (42 new), update-checker-github.test.ts deleted
(superceded by tests/upgrade/checker.test.ts).
