+++ packages/myco/src/daemon/update-checker.ts
@@ -25,9 +25,7 @@ import { loadMachineConfig, updateTierConfigRaw } from '../config/loader.js';
 import { setAtPath } from '../utils/dot-path.js';
 
 import {
-  NPM_REGISTRY_BASE_URL,
   NPM_PACKAGE_NAME,
-  UPDATE_PACKAGES,
   MYCO_GLOBAL_DIR,
   UPDATE_CHECK_CACHE_PATH,
   UPDATE_CONFIG_PATH,
@@ -47,11 +45,6 @@ import {
 import { clearJsonSentinel } from '../utils/json-sentinel.js';
 import { readJsonFile } from '../utils/json.js';
 import { getPluginVersion } from '../version.js';
-import {
-  mycoReleasesApiUrl,
-  resolveMycoVersions,
-  githubHeaders,
-} from '@myco/upgrade/release-assets.js';
 
 // ---------------------------------------------------------------------------
 // Public types
@@ -77,60 +70,9 @@ export interface CachedCheck {
   packages: Partial<Record<UpdatePackageId, CachedPackageCheck>>;
 }
 
-/** Installed/update status for one globally installed Myco package. */
-export interface PackageCheckResult {
-  id: UpdatePackageId;
-  display_name: string;
-  package_name: string;
-  installed: boolean;
-  installed_version: string | null;
-  latest_version: string | null;
-  latest_stable: string | null;
-  latest_beta: string | null;
-  /** True when the target version is strictly greater than what's installed. */
-  update_available: boolean;
-  /**
-   * True when the desired channel is `stable`, the running binary is a
-   * prerelease, and a stable target exists — i.e. a beta user can step back
-   * onto the stable release. Set only on the `myco` package; mutually
-   * exclusive with `update_available`. Derived from the running version +
-   * desired channel (the robust primary signal), corroborated by the install
-   * marker when present — NOT from the retired managed-runtime pin.
-   */
-  revert_available: boolean;
-}
-
-/** Result returned to callers of checkForUpdate / statusFromCache */
-export interface CheckResult {
-  update_available: boolean;
-  /** True when any package has `revert_available` set. */
-  revert_available: boolean;
-  running_version: string;
-  latest_version: string;
-  latest_stable: string;
-  latest_beta: string | null;
-  channel: ReleaseChannel;
-  /** Always `'machine'` — the channel is machine-scoped (decision-46130740). */
-  channel_scope: 'machine';
-  /**
-   * Always `'machine'` — myco is a single managed binary at `~/.myco/bin/myco`
-   * swapped in place (curl + npm both install it). The legacy `'managed'`
-   * value (a separate `~/.myco/runtime/` npm install) was retired with the
-   * native installer; the field is kept in the API contract for stability.
-   */
-  runtime_scope: 'machine';
-  check_interval_hours: number;
-  last_check: string;
-  error: string | null;
-  packages: PackageCheckResult[];
-}
-
 // ---------------------------------------------------------------------------
 // Constants
 // ---------------------------------------------------------------------------
-
-/** Fetch timeout for registry requests. */
-const REGISTRY_FETCH_TIMEOUT_MS = 10_000;
 
 // ---------------------------------------------------------------------------
 // Dev-mode exemption
@@ -649,175 +591,6 @@ export function consumeUpdateError(): void {
 }
 
 // ---------------------------------------------------------------------------
-// Registry types
-// ---------------------------------------------------------------------------
-
-interface NpmDistTags {
-  latest: string;
-  beta?: string;
-  [tag: string]: string | undefined;
-}
-
-interface NpmRegistryResponse {
-  'dist-tags': NpmDistTags;
-}
-
-/** Build the npm registry URL for a specific package. */
-function packageRegistryUrl(packageName: string): string {
-  return `${NPM_REGISTRY_BASE_URL}/${encodeURIComponent(packageName)}`;
-}
-
-// ---------------------------------------------------------------------------
-// Channel comparison logic
-// ---------------------------------------------------------------------------
-
-/**
- * Returns the target version to compare against based on channel.
- * - Stable: dist-tags.latest
- * - Beta: max(dist-tags.latest, dist-tags.beta) — no-downgrade rule
- */
-function resolveTargetVersion(distTags: NpmDistTags, channel: ReleaseChannel): string {
-  const stable = distTags.latest;
-  const beta = distTags.beta ?? null;
-
-  if (channel === 'stable' || beta === null) {
-    return stable;
-  }
-
-  // Beta channel: pick whichever is higher (stable can exceed beta tag)
-  const higher = semver.gt(beta, stable) ? beta : stable;
-  return higher;
-}
-
-function resolveTargetVersionFromCache(
-  pkg: CachedPackageCheck,
-  channel: ReleaseChannel,
-): string {
-  return resolveTargetVersion(
-    { latest: pkg.latest_stable, beta: pkg.latest_beta ?? undefined },
-    channel,
-  );
-}
-
-function buildInstalledPackageVersions(
-  globalPrefix: string | null,
-  currentVersion: string,
-): Record<UpdatePackageId, string | null> {
-  const installed: Record<UpdatePackageId, string | null> = {
-    myco: currentVersion,
-    'myco-team': null,
-    'myco-collective': null,
-  };
-
-  if (globalPrefix === null) return installed;
-
-  for (const pkg of UPDATE_PACKAGES) {
-    if (pkg.id === 'myco') continue;
-    installed[pkg.id] = getInstalledVersion(globalPrefix, pkg.packageName);
-  }
-
-  return installed;
-}
-
-/**
- * True when `version` carries a semver prerelease component (e.g. a `-beta.N`
- * suffix). The robust primary signal that the running myco binary is a beta:
- * curl + npm both write the same `~/.myco/bin/myco`, so the running version is
- * authoritative where an install-marker channel may lag.
- */
-function isPrerelease(version: string): boolean {
-  return semver.valid(version) !== null && semver.prerelease(version) !== null;
-}
-
-
-function buildPackageResults(
-  currentVersion: string,
-  cache: CachedCheck,
-  channel: ReleaseChannel,
-  globalPrefix: string | null,
-): PackageCheckResult[] {
-  const installedVersions = buildInstalledPackageVersions(globalPrefix, currentVersion);
-  // Revert-to-stable is offered when the operator's DESIRED channel is stable
-  // but the running binary is a prerelease. The running version is the
-  // authoritative signal: the managed-binary swap (curl + npm) does not rewrite
-  // the install marker, so the marker may show 'stable' while the binary is
-  // already a prerelease. Keying on the running version makes revert available
-  // in all cases, including that stale-marker scenario.
-  const desiredStableRevert =
-    channel === 'stable'
-    && isPrerelease(currentVersion);
-
-  return UPDATE_PACKAGES.map((pkg) => {
-    const cached = cache.packages[pkg.id];
-    const installedVersion = installedVersions[pkg.id];
-    const latestVersion = cached ? resolveTargetVersionFromCache(cached, channel) : null;
-    const updateAvailable =
-      installedVersion !== null &&
-      latestVersion !== null &&
-      semver.valid(installedVersion) !== null &&
-      semver.valid(latestVersion) !== null &&
-      semver.gt(latestVersion, installedVersion);
-    const revertAvailable =
-      pkg.id === 'myco' &&
-      desiredStableRevert &&
-      cached?.latest_stable != null &&
-      latestVersion !== null &&
-      latestVersion !== currentVersion &&
-      !updateAvailable;
-
-    return {
-      id: pkg.id,
-      display_name: pkg.displayName,
-      package_name: pkg.packageName,
-      installed: installedVersion !== null,
-      installed_version: installedVersion,
-      latest_version: latestVersion,
-      latest_stable: cached?.latest_stable ?? null,
-      latest_beta: cached?.latest_beta ?? null,
-      update_available: updateAvailable,
-      revert_available: revertAvailable,
-    };
-  });
-}
-
-// ---------------------------------------------------------------------------
-// CheckResult builder
-// ---------------------------------------------------------------------------
-
-function buildCheckResult(
-  currentVersion: string,
-  cache: CachedCheck,
-  config: UpdateConfig,
-  channel: ReleaseChannel,
-  error: string | null,
-  globalPrefix: string | null,
-): CheckResult {
-  const packages = buildPackageResults(currentVersion, cache, channel, globalPrefix);
-  const primaryPackage = packages.find((pkg) => pkg.id === 'myco');
-  const targetVersion = primaryPackage?.latest_version ?? currentVersion;
-  const latestStable = primaryPackage?.latest_stable ?? currentVersion;
-  const latestBeta = primaryPackage?.latest_beta ?? null;
-  const updateAvailable = packages.some((pkg) => pkg.installed && pkg.update_available);
-  const revertAvailable = packages.some((pkg) => pkg.revert_available);
-
-  return {
-    update_available: updateAvailable,
-    revert_available: revertAvailable,
-    running_version: currentVersion,
-    latest_version: targetVersion,
-    latest_stable: latestStable,
-    latest_beta: latestBeta,
-    channel,
-    channel_scope: 'machine',
-    runtime_scope: 'machine',
-    check_interval_hours: config.check_interval_hours,
-    last_check: cache.checked_at,
-    error,
-    packages,
-  };
-}
-
-// ---------------------------------------------------------------------------
 // Installed version detection
 // ---------------------------------------------------------------------------
 
@@ -855,168 +628,3 @@ export function getInstalledVersion(
   }
 }
 
-// ---------------------------------------------------------------------------
-// Primary exports
-// ---------------------------------------------------------------------------
-
-/**
- * Fetches the npm registry, compares versions, and writes the result to cache.
- *
- * On network failure, returns the last cached result (with an error field) if
- * one exists. If no cache exists and the fetch fails, the error field is set
- * and update_available is false.
- */
-export async function checkForUpdate(
-  currentVersion: string,
-  globalPrefix: string | null = null,
-  runtimeCommand: string | null = null,
-  channelOverride?: ReleaseChannel,
-): Promise<CheckResult> {
-  const config = readUpdateConfig();
-  const existingCache = readCachedCheck();
-  const effectiveChannel = channelOverride ?? config.channel;
-
-  const freshPackages: Partial<Record<UpdatePackageId, CachedPackageCheck>> = {};
-  const fetchErrors: string[] = [];
-
-  const registryChecks = await Promise.allSettled(
-    UPDATE_PACKAGES.map(async (pkg) => {
-      if (pkg.id === 'myco') {
-        const response = await fetch(mycoReleasesApiUrl(), {
-          headers: githubHeaders(),
-          signal: AbortSignal.timeout(REGISTRY_FETCH_TIMEOUT_MS),
-        });
-
-        if (!response.ok) {
-          throw new Error(`${pkg.packageName}: GitHub releases responded with ${response.status}`);
-        }
-
-        const releases = await response.json();
-        const { latest_stable, latest_beta } = resolveMycoVersions(releases);
-        return {
-          id: pkg.id,
-          package_name: pkg.packageName,
-          latest_stable: latest_stable ?? currentVersion,
-          latest_beta,
-        };
-      }
-
-      const response = await fetch(packageRegistryUrl(pkg.packageName), {
-        signal: AbortSignal.timeout(REGISTRY_FETCH_TIMEOUT_MS),
-      });
-
-      if (!response.ok) {
-        throw new Error(`${pkg.packageName}: registry responded with ${response.status}`);
-      }
-
-      const data = (await response.json()) as NpmRegistryResponse;
-      return {
-        id: pkg.id,
-        package_name: pkg.packageName,
-        latest_stable: data['dist-tags'].latest,
-        latest_beta: data['dist-tags'].beta ?? null,
-      };
-    }),
-  );
-
-  for (const result of registryChecks) {
-    if (result.status === 'fulfilled') {
-      freshPackages[result.value.id] = {
-        package_name: result.value.package_name,
-        latest_stable: result.value.latest_stable,
-        latest_beta: result.value.latest_beta,
-      };
-      continue;
-    }
-
-    const message = result.reason instanceof Error ? result.reason.message : String(result.reason);
-    fetchErrors.push(message);
-  }
-
-  if (existingCache !== null) {
-    for (const pkg of UPDATE_PACKAGES) {
-      if (freshPackages[pkg.id] !== undefined) continue;
-      const cached = existingCache.packages[pkg.id];
-      if (cached) {
-        freshPackages[pkg.id] = cached;
-      }
-    }
-  }
-
-  if (Object.keys(freshPackages).length === 0) {
-    const fetchError = fetchErrors[0] ?? 'registry fetch failed';
-    return {
-      update_available: false,
-      revert_available: false,
-      running_version: currentVersion,
-      latest_version: currentVersion,
-      latest_stable: currentVersion,
-      latest_beta: null,
-      channel: effectiveChannel,
-      channel_scope: 'machine',
-      runtime_scope: 'machine',
-      check_interval_hours: config.check_interval_hours,
-      last_check: new Date().toISOString(),
-      error: fetchError,
-      packages: buildPackageResults(
-        currentVersion,
-        { checked_at: new Date().toISOString(), channel: effectiveChannel, packages: {} },
-        effectiveChannel,
-        globalPrefix,
-      ),
-    };
-  }
-
-  const freshCache: CachedCheck = {
-    checked_at: new Date().toISOString(),
-    channel: effectiveChannel,
-    packages: freshPackages,
-  };
-
-  try {
-    fs.mkdirSync(path.dirname(UPDATE_CHECK_CACHE_PATH), { recursive: true });
-    fs.writeFileSync(UPDATE_CHECK_CACHE_PATH, JSON.stringify(freshCache, null, 2), 'utf-8');
-  } catch {
-    // Cache write failure is non-fatal
-  }
-
-  const error = fetchErrors.length > 0 ? fetchErrors.join('; ') : null;
-  return buildCheckResult(
-    currentVersion,
-    freshCache,
-    config,
-    effectiveChannel,
-    error,
-    globalPrefix,
-  );
-}
-
-/**
- * Builds a CheckResult from cached data without hitting the registry.
- * Returns null when no cache exists.
- *
- * Accepts optional pre-read `cache` and `config` to avoid redundant file
- * reads when the caller has already loaded them (e.g. for a staleness check).
- */
-export function statusFromCache(
-  currentVersion: string,
-  cache?: CachedCheck | null,
-  config?: UpdateConfig,
-  globalPrefix: string | null = null,
-  runtimeCommand: string | null = null,
-  channelOverride?: ReleaseChannel,
-): CheckResult | null {
-  const resolvedCache = cache !== undefined ? cache : readCachedCheck();
-  if (resolvedCache === null) return null;
-
-  const resolvedConfig = config !== undefined ? config : readUpdateConfig();
-  const effectiveChannel = channelOverride ?? resolvedCache.channel ?? resolvedConfig.channel;
-  return buildCheckResult(
-    currentVersion,
-    resolvedCache,
-    resolvedConfig,
-    effectiveChannel,
-    null,
-    globalPrefix,
-  );
-}
