--- tests/daemon/api/update.test.ts
@@ -1,981 +1,0 @@
-/**
- * Tests for the update API route handlers.
- *
- * Covers:
- * - handleUpdateStatus: exempt, fresh cache (no background check), stale cache (background check)
- * - handleUpdateCheck: forces registry fetch, 400 when exempt
- * - handleUpdateApply: spawns script + schedules shutdown, 400 when no update, 400 when exempt
- * - handleUpdateChannel: writes config + clears cache, 400 for invalid channel
- */
-
-import { describe, it, expect, beforeEach, afterEach, afterAll, mock } from 'bun:test';
-import { vi } from '../../helpers/vi-shim.js';
-import fs from 'node:fs';
-import os from 'node:os';
-import path from 'node:path';
-// ---------------------------------------------------------------------------
-// Module mocks — hoisted before imports
-// ---------------------------------------------------------------------------
-
-// Capture the real modules BEFORE replacing them, so afterAll can put them
-// back. bun:test's mock.module() is process-scoped — without this restore,
-// the stub modules below leak into every later test file that imports them
-// in the same `bun test` invocation (the canonical npm test runner uses
-// --isolate, which masks the leak; the restore stays correct regardless).
-const realUpdateChecker = await import('@myco/daemon/update-checker.js');
-const realUpdateInstaller = await import('@myco/upgrade/spawn.js');
-const realReleaseResolver = await import('@myco/upgrade/release-resolver.js');
-
-mock.module('@myco/daemon/update-checker.js', () => ({
-  isUpdateExempt: vi.fn(() => false),
-  checkForUpdate: vi.fn(),
-  statusFromCache: vi.fn(),
-  readCachedCheck: vi.fn(() => null),
-  readUpdateConfig: vi.fn(() => ({ channel: 'stable', check_interval_hours: 6 })),
-  readProjectReleaseChannel: vi.fn(() => 'stable'),
-  writeProjectReleaseChannel: vi.fn(),
-  clearCachedCheck: vi.fn(),
-  isCacheStale: vi.fn(() => false),
-  getInstalledVersion: vi.fn(() => null),
-  resolveMycoBinary: vi.fn(() => 'myco'),
-  resolveRuntimeCommand: vi.fn(() => null),
-}));
-
-mock.module('@myco/upgrade/spawn.js', () => ({
-  spawnUpdateScript: vi.fn(() => '/tmp/myco-update-123.sh'),
-  spawnRestartScript: vi.fn(() => '/tmp/myco-restart-123.sh'),
-}));
-
-// The myco binary path resolves release refs through this module. Mock it so
-// handleUpdateApply never hits the network; tests drive the resolved refs.
-mock.module('@myco/upgrade/release-resolver.js', () => ({
-  resolveMycoBinaryUpdateRefs: vi.fn(),
-  resolveMycoBinaryUpdateRefsForVersion: vi.fn(),
-}));
-
-afterAll(() => {
-  mock.module('@myco/daemon/update-checker.js', () => realUpdateChecker);
-  mock.module('@myco/upgrade/spawn.js', () => realUpdateInstaller);
-  mock.module('@myco/upgrade/release-resolver.js', () => realReleaseResolver);
-});
-
-import {
-  isUpdateExempt,
-  checkForUpdate,
-  statusFromCache,
-  readCachedCheck,
-  readUpdateConfig,
-  readProjectReleaseChannel,
-  writeProjectReleaseChannel,
-  clearCachedCheck,
-  isCacheStale,
-  getInstalledVersion,
-  resolveRuntimeCommand,
-} from '@myco/daemon/update-checker.js';
-import { spawnUpdateScript, spawnRestartScript } from '@myco/upgrade/spawn.js';
-import { resolveMycoBinaryUpdateRefs } from '@myco/upgrade/release-resolver.js';
-import { createUpdateHandlers } from '@myco/daemon/api/update.js';
-import type { RouteRequest } from '@myco/daemon/router.js';
-import { FakeServiceManager } from '../../helpers/fake-service-manager';
-
-// ---------------------------------------------------------------------------
-// Pre-installed service helper. The /update routes look up
-// restartShellCommand only when the service is already installed for this
-// PID — wire a status snapshot + shell command in one call so each test
-// reads as a single intent line.
-// ---------------------------------------------------------------------------
-function installedServiceManager(label: string, shellCmd: string): FakeServiceManager {
-  const mgr = new FakeServiceManager();
-  mgr.installed.add(label);
-  mgr.statuses.set(label, { installed: true, running: true, pid: process.pid, lastExitCode: 0, unitPath: '/x' });
-  mgr.restartShellCommands.set(label, shellCmd);
-  return mgr;
-}
-
-import { TEST_REQUEST_CONTEXT } from '../../helpers/request-context';
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/** Minimal stub RouteRequest. */
-function makeReq(overrides: Partial<RouteRequest> = {}): RouteRequest {
-  return {
-    body: {},
-    requestContext: TEST_REQUEST_CONTEXT,
-    query: {},
-    params: {},
-    pathname: '/api/update/status',
-    ...overrides,
-  };
-}
-
-/** A representative CheckResult with update available. */
-const UPDATE_AVAILABLE_STATUS = {
-  update_available: true,
-  running_version: '1.0.0',
-  latest_version: '1.1.0',
-  latest_stable: '1.1.0',
-  latest_beta: null,
-  channel: 'stable' as const,
-  check_interval_hours: 6,
-  last_check: '2026-03-28T00:00:00.000Z',
-  error: null,
-  packages: [
-    {
-      id: 'myco',
-      display_name: 'Myco',
-      package_name: '@goondocks/myco',
-      installed: true,
-      installed_version: '1.0.0',
-      latest_version: '1.1.0',
-      latest_stable: '1.1.0',
-      latest_beta: null,
-      update_available: true,
-    },
-  ],
-};
-
-/** A representative CheckResult with no update available. */
-const NO_UPDATE_STATUS = {
-  ...UPDATE_AVAILABLE_STATUS,
-  update_available: false,
-  latest_version: '1.0.0',
-  latest_stable: '1.0.0',
-  packages: [
-    {
-      ...UPDATE_AVAILABLE_STATUS.packages[0],
-      latest_version: '1.0.0',
-      latest_stable: '1.0.0',
-      update_available: false,
-    },
-  ],
-};
-
-/** Resolved myco binary-update refs the release resolver returns for 1.1.0. */
-const MYCO_REFS_1_1_0 = {
-  assetUrl: 'https://example.test/releases/myco-darwin-arm64',
-  sha256sumsUrl: 'https://example.test/releases/SHA256SUMS',
-  assetName: 'myco-darwin-arm64',
-  targetVersion: '1.1.0',
-};
-
-/** Resolved myco binary-update refs for the revert-to-stable 1.0.0 target. */
-const MYCO_REFS_1_0_0 = {
-  ...MYCO_REFS_1_1_0,
-  targetVersion: '1.0.0',
-};
-
-/** Default deps for tests. Each call gets a fresh tmpdir for the
- *  update-in-progress sentinel so tests don't interfere across runs. */
-function makeDeps(overrides: Record<string, unknown> = {}) {
-  const daemonStateDir = fs.mkdtempSync(path.join(os.tmpdir(), 'myco-update-test-'));
-  return {
-    vaultDir: '/vault',
-    projectRoot: '/project',
-    currentVersion: '1.0.0',
-    daemonPort: 20915,
-    daemonStateDir,
-    scheduleShutdown: vi.fn(),
-    ...overrides,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// handleUpdateStatus
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateStatus', () => {
-  beforeEach(() => {
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readCachedCheck).mockReturnValue(null);
-    vi.mocked(readUpdateConfig).mockReturnValue({ channel: 'stable', check_interval_hours: 6 });
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(isCacheStale).mockReturnValue(false);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    vi.mocked(checkForUpdate).mockResolvedValue(NO_UPDATE_STATUS);
-  });
-
-  it('returns exempt:true when in dev mode', async () => {
-    vi.mocked(isUpdateExempt).mockReturnValue(true);
-    const { handleUpdateStatus } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(result.status).toBeUndefined();
-    expect(result.body).toEqual({ exempt: true, running_version: '1.0.0' });
-  });
-
-  it('returns status from cache when cache is fresh', async () => {
-    const freshCache = {
-      checked_at: new Date().toISOString(),
-      current_version: '1.0.0',
-      latest_stable: '1.0.0',
-      latest_beta: null,
-      channel: 'stable' as const,
-    };
-    vi.mocked(readCachedCheck).mockReturnValue(freshCache);
-    vi.mocked(isCacheStale).mockReturnValue(false);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-
-    const { handleUpdateStatus } = createUpdateHandlers(makeDeps());
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(result.body).toMatchObject({ exempt: false, update_available: false });
-    // Should NOT trigger a background check
-    expect(checkForUpdate).not.toHaveBeenCalled();
-  });
-
-  it('kicks off background check when cache is stale', async () => {
-    vi.mocked(isCacheStale).mockReturnValue(true);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    vi.mocked(checkForUpdate).mockResolvedValue(NO_UPDATE_STATUS);
-
-    const { handleUpdateStatus } = createUpdateHandlers(makeDeps());
-    const result = await handleUpdateStatus(makeReq());
-
-    // Response returned immediately (does not await checkForUpdate)
-    expect(result.body).toMatchObject({ exempt: false });
-    // Background check was triggered
-    expect(checkForUpdate).toHaveBeenCalledWith('1.0.0', undefined, null, 'stable');
-  });
-
-  it('returns exempt:false in body when not exempt', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    const { handleUpdateStatus } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect((result.body as Record<string, unknown>).exempt).toBe(false);
-  });
-
-  it('returns default status when cache is empty (null)', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(null);
-    vi.mocked(isCacheStale).mockReturnValue(true);
-    vi.mocked(checkForUpdate).mockResolvedValue(NO_UPDATE_STATUS);
-
-    const { handleUpdateStatus } = createUpdateHandlers(makeDeps());
-    const result = await handleUpdateStatus(makeReq());
-    const body = result.body as Record<string, unknown>;
-
-    expect(body.exempt).toBe(false);
-    expect(body.update_available).toBe(false);
-    expect(body.running_version).toBe('1.0.0');
-    expect(body.channel).toBe('stable');
-    expect(body.channel_scope).toBe('machine');
-    expect(body.runtime_scope).toBe('machine');
-    expect(body.last_check).toBe('');
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleUpdateCheck
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateCheck', () => {
-  beforeEach(() => {
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(checkForUpdate).mockResolvedValue(UPDATE_AVAILABLE_STATUS);
-  });
-
-  it('returns 400 when exempt', async () => {
-    vi.mocked(isUpdateExempt).mockReturnValue(true);
-    const { handleUpdateCheck } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateCheck(makeReq());
-
-    expect(result.status).toBe(400);
-    expect((result.body as Record<string, unknown>).error).toBe('update_exempt');
-  });
-
-  it('awaits checkForUpdate and returns result', async () => {
-    const { handleUpdateCheck } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateCheck(makeReq());
-
-    expect(checkForUpdate).toHaveBeenCalledWith('1.0.0', undefined, null, 'stable');
-    expect(result.body).toMatchObject({ exempt: false, update_available: true });
-  });
-
-  it('propagates update_available:false when already up to date', async () => {
-    vi.mocked(checkForUpdate).mockResolvedValue(NO_UPDATE_STATUS);
-    const { handleUpdateCheck } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateCheck(makeReq());
-
-    expect((result.body as Record<string, unknown>).update_available).toBe(false);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleUpdateApply
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateApply', () => {
-  beforeEach(() => {
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(statusFromCache).mockReturnValue(UPDATE_AVAILABLE_STATUS);
-    vi.mocked(spawnUpdateScript).mockReset();
-    vi.mocked(spawnUpdateScript).mockReturnValue('/tmp/myco-update-123.sh');
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    // Default: the release resolver yields the 1.1.0 myco refs. Tests that
-    // need a different target (revert) or a no-myco-update case override this.
-    vi.mocked(resolveMycoBinaryUpdateRefs).mockReset();
-    vi.mocked(resolveMycoBinaryUpdateRefs).mockResolvedValue(MYCO_REFS_1_1_0);
-  });
-
-  it('returns 400 when exempt', async () => {
-    vi.mocked(isUpdateExempt).mockReturnValue(true);
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateApply(makeReq());
-
-    expect(result.status).toBe(400);
-    expect((result.body as Record<string, unknown>).error).toBe('update_exempt');
-  });
-
-  it('returns 400 when no update is available', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateApply(makeReq());
-
-    expect(result.status).toBe(400);
-    expect((result.body as Record<string, unknown>).error).toBe('no_update_available');
-  });
-
-  it('returns 400 when cache is empty (no status)', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(null);
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateApply(makeReq());
-
-    expect(result.status).toBe(400);
-  });
-
-  it('routes myco through the BINARY swap (no myco in packageSpecs) and schedules shutdown', async () => {
-    const scheduleShutdown = vi.fn();
-    const deps = makeDeps({ scheduleShutdown });
-    const { handleUpdateApply } = createUpdateHandlers(deps);
-
-    const result = await handleUpdateApply(makeReq());
-
-    // Myco is NOT in packageSpecs — it's the binary swap. Operator CLIs (none
-    // here) would be the only packageSpecs entries.
-    expect(spawnUpdateScript).toHaveBeenCalledWith({
-      packageSpecs: [],
-      projectRoot: '/project',
-      vaultDir: '/vault',
-      mycoBinary: 'myco',
-      serviceManagedLabel: null,
-      daemonPort: 20915,
-      inProgressSentinelPath: path.join(deps.daemonStateDir, 'update.in-progress'),
-      targetVersion: '1.1.0',
-      mycoBinaryUpdate: MYCO_REFS_1_1_0,
-    });
-    // Resolver was consulted for the stable channel.
-    expect(resolveMycoBinaryUpdateRefs).toHaveBeenCalledWith('stable');
-    expect(scheduleShutdown).toHaveBeenCalled();
-    expect(result.body).toMatchObject({ status: 'applying', version: '1.1.0' });
-  });
-
-  it('does not spawn script when update is unavailable', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps());
-
-    await handleUpdateApply(makeReq());
-
-    expect(spawnUpdateScript).not.toHaveBeenCalled();
-  });
-
-  it('keeps operator CLIs (myco-team) on npm while myco takes the binary swap', async () => {
-    vi.mocked(statusFromCache).mockReturnValue({
-      ...UPDATE_AVAILABLE_STATUS,
-      packages: [
-        ...UPDATE_AVAILABLE_STATUS.packages,
-        {
-          id: 'myco-team',
-          display_name: 'Myco Team',
-          package_name: '@goondocks/myco-team',
-          installed: true,
-          installed_version: '0.1.0',
-          latest_version: '0.1.1',
-          latest_stable: '0.1.1',
-          latest_beta: null,
-          update_available: true,
-        },
-      ],
-    });
-    const deps = makeDeps();
-    const { handleUpdateApply } = createUpdateHandlers(deps);
-
-    await handleUpdateApply(makeReq());
-
-    expect(spawnUpdateScript).toHaveBeenCalledWith({
-      // Operator CLI STAYS on npm; myco is NOT in packageSpecs.
-      packageSpecs: ['@goondocks/myco-team@0.1.1'],
-      projectRoot: '/project',
-      vaultDir: '/vault',
-      mycoBinary: 'myco',
-      serviceManagedLabel: null,
-      daemonPort: 20915,
-      inProgressSentinelPath: path.join(deps.daemonStateDir, 'update.in-progress'),
-      targetVersion: '1.1.0',
-      mycoBinaryUpdate: MYCO_REFS_1_1_0,
-    });
-  });
-
-  it('revert-to-stable (beta-running, desired stable) resolves the STABLE release and swaps the same binary — no 400, no npm', async () => {
-    // A beta user with desired channel stable. revert_available is true and
-    // update_available is false (it's a downgrade), so WITHOUT enteringStable
-    // this 400s `no_update_available`. enteringStable (re-founded on the
-    // running prerelease version, not a managed-runtime pin) resolves the
-    // stable release and swaps the SAME ~/.myco/bin/myco.
-    vi.mocked(statusFromCache).mockReturnValue({
-      ...UPDATE_AVAILABLE_STATUS,
-      running_version: '1.1.0-beta.1',
-      latest_version: '1.0.0',
-      latest_stable: '1.0.0',
-      update_available: false,
-      revert_available: true,
-      packages: [
-        {
-          ...UPDATE_AVAILABLE_STATUS.packages[0],
-          installed_version: '1.1.0-beta.1',
-          latest_version: '1.0.0',
-          latest_stable: '1.0.0',
-          update_available: false,
-          revert_available: true,
-        },
-      ],
-    });
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    // The stable channel resolves the 1.0.0 release for the revert.
-    vi.mocked(resolveMycoBinaryUpdateRefs).mockResolvedValue(MYCO_REFS_1_0_0);
-
-    // Daemon is RUNNING the beta prerelease — this is the enteringStable signal.
-    const deps = makeDeps({ currentVersion: '1.1.0-beta.1' });
-    const { handleUpdateApply } = createUpdateHandlers(deps);
-
-    const result = await handleUpdateApply(makeReq());
-
-    // NOT a 400.
-    expect(result.status).toBeUndefined();
-    expect(spawnUpdateScript).toHaveBeenCalledWith({
-      // No myco npm spec, no managed-runtime plumbing — the binary swaps to 1.0.0.
-      packageSpecs: [],
-      projectRoot: '/project',
-      vaultDir: '/vault',
-      mycoBinary: 'myco',
-      serviceManagedLabel: null,
-      daemonPort: 20915,
-      inProgressSentinelPath: path.join(deps.daemonStateDir, 'update.in-progress'),
-      targetVersion: '1.0.0',
-      mycoBinaryUpdate: MYCO_REFS_1_0_0,
-    });
-    // The revert resolved through the stable channel.
-    expect(resolveMycoBinaryUpdateRefs).toHaveBeenCalledWith('stable');
-  });
-
-  it('enteringStable: beta-running + desired stable swaps to stable even when revert_available is somehow stale', async () => {
-    // Guard the symmetric apply path independently of revert_available: a
-    // prerelease binary on the stable channel must always be offered the swap.
-    vi.mocked(statusFromCache).mockReturnValue({
-      ...UPDATE_AVAILABLE_STATUS,
-      channel: 'stable',
-      running_version: '1.1.0-beta.1',
-      latest_version: '1.0.0',
-      latest_stable: '1.0.0',
-      update_available: false,
-      revert_available: false,
-      packages: [
-        {
-          ...UPDATE_AVAILABLE_STATUS.packages[0],
-          installed_version: '1.1.0-beta.1',
-          latest_version: '1.0.0',
-          latest_stable: '1.0.0',
-          update_available: false,
-          revert_available: false,
-        },
-      ],
-    });
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    vi.mocked(resolveMycoBinaryUpdateRefs).mockResolvedValue(MYCO_REFS_1_0_0);
-
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ currentVersion: '1.1.0-beta.1' }));
-
-    const result = await handleUpdateApply(makeReq());
-
-    expect(result.status).toBeUndefined();
-    expect(resolveMycoBinaryUpdateRefs).toHaveBeenCalledWith('stable');
-    expect(vi.mocked(spawnUpdateScript).mock.calls[0][0]).toMatchObject({
-      packageSpecs: [],
-      targetVersion: '1.0.0',
-      mycoBinaryUpdate: MYCO_REFS_1_0_0,
-    });
-  });
-
-  // -------------------------------------------------------------------------
-  // Service-managed restart routing
-  //
-  // Mirrors the bug fix in commit 78a2c421 for /restart. When launchd /
-  // systemd manages the daemon, the orchestrator must NOT spawn `myco daemon`
-  // itself — the service supervisor's KeepAlive would race it for the
-  // canonical port. Instead, the orchestrator restarts via the platform
-  // ServiceManager. The daemon passes the detected service LABEL (not a shell
-  // command); the cross-platform orchestrator turns the label into a
-  // ServiceManager.restart() call.
-  // -------------------------------------------------------------------------
-  it('passes the service label when the daemon is service-managed (prod)', async () => {
-    const mgr = installedServiceManager(
-      'co.goondocks.myco',
-      'launchctl kickstart -k gui/501/co.goondocks.myco',
-    );
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ serviceManager: mgr }));
-
-    await handleUpdateApply(makeReq());
-
-    expect(spawnUpdateScript).toHaveBeenCalledWith(
-      expect.objectContaining({
-        serviceManagedLabel: 'co.goondocks.myco',
-      }),
-    );
-  });
-
-  it('passes the dev service label for the dev service variant', async () => {
-    const mgr = installedServiceManager(
-      'co.goondocks.myco-dev',
-      'launchctl kickstart -k gui/501/co.goondocks.myco-dev',
-    );
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ serviceManager: mgr }));
-
-    await handleUpdateApply(makeReq());
-
-    expect(spawnUpdateScript).toHaveBeenCalledWith(
-      expect.objectContaining({
-        serviceManagedLabel: 'co.goondocks.myco-dev',
-      }),
-    );
-  });
-
-  it('passes a null label when no service is installed (manual daemon)', async () => {
-    const mgr = new FakeServiceManager();
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ serviceManager: mgr }));
-
-    await handleUpdateApply(makeReq());
-
-    const call = vi.mocked(spawnUpdateScript).mock.calls[0]?.[0];
-    expect(call?.serviceManagedLabel).toBeNull();
-  });
-
-  it('passes a null label when service is installed but a different PID is the daemon', async () => {
-    // The supervisor manages SOME process, but not this one. Don't claim
-    // service-managed semantics — fall back to the manual daemon spawn.
-    const mgr = new FakeServiceManager();
-    mgr.installed.add('co.goondocks.myco');
-    mgr.statuses.set('co.goondocks.myco', { installed: true, running: true, pid: process.pid + 999, lastExitCode: 0, unitPath: '/x' });
-    mgr.restartShellCommands.set('co.goondocks.myco', 'launchctl kickstart -k gui/501/co.goondocks.myco');
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ serviceManager: mgr }));
-
-    await handleUpdateApply(makeReq());
-
-    const call = vi.mocked(spawnUpdateScript).mock.calls[0]?.[0];
-    expect(call?.serviceManagedLabel).toBeNull();
-  });
-
-  it('passes a null label on unsupported platforms', async () => {
-    const mgr = new FakeServiceManager({ supported: false });
-    const { handleUpdateApply } = createUpdateHandlers(makeDeps({ serviceManager: mgr }));
-
-    await handleUpdateApply(makeReq());
-
-    const call = vi.mocked(spawnUpdateScript).mock.calls[0]?.[0];
-    expect(call?.serviceManagedLabel).toBeNull();
-  });
-
-  it('entering beta swaps to the beta binary even when the version matches (resolves the prerelease)', async () => {
-    // User is on stable, opts into beta. The running version equals the beta
-    // target's base — update_available is false — but the channel switch still
-    // swaps onto the beta binary (the old "create managed runtime" semantic,
-    // now a binary swap).
-    vi.mocked(statusFromCache).mockReturnValue({
-      ...UPDATE_AVAILABLE_STATUS,
-      channel: 'beta',
-      update_available: false,
-      revert_available: false,
-      running_version: '1.0.0',
-      latest_version: '1.0.0',
-      packages: [
-        {
-          ...UPDATE_AVAILABLE_STATUS.packages[0],
-          installed_version: '1.0.0',
-          latest_version: '1.0.0',
-          update_available: false,
-          revert_available: false,
-        },
-      ],
-    });
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    const betaRefs = { ...MYCO_REFS_1_1_0, targetVersion: '1.0.0' };
-    vi.mocked(resolveMycoBinaryUpdateRefs).mockResolvedValue(betaRefs);
-
-    const deps = makeDeps();
-    const { handleUpdateApply } = createUpdateHandlers(deps);
-
-    const result = await handleUpdateApply(makeReq());
-
-    expect(result.status).toBeUndefined();
-    // Beta channel was resolved.
-    expect(resolveMycoBinaryUpdateRefs).toHaveBeenCalledWith('beta');
-    expect(spawnUpdateScript).toHaveBeenCalledWith({
-      packageSpecs: [],
-      projectRoot: '/project',
-      vaultDir: '/vault',
-      mycoBinary: 'myco',
-      serviceManagedLabel: null,
-      daemonPort: 20915,
-      inProgressSentinelPath: path.join(deps.daemonStateDir, 'update.in-progress'),
-      targetVersion: '1.0.0',
-      mycoBinaryUpdate: betaRefs,
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleUpdateChannel
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateChannel', () => {
-  beforeEach(() => {
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readUpdateConfig).mockReturnValue({ channel: 'stable', check_interval_hours: 6 });
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    vi.mocked(writeProjectReleaseChannel).mockImplementation(() => {});
-    vi.mocked(clearCachedCheck).mockImplementation(() => {});
-  });
-
-  it('returns 400 for an invalid channel', async () => {
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateChannel(makeReq({ body: { channel: 'nightly' } }));
-
-    expect(result.status).toBe(400);
-    expect((result.body as Record<string, unknown>).error).toBe('invalid_channel');
-  });
-
-  it('returns 400 when channel is missing', async () => {
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateChannel(makeReq({ body: {} }));
-
-    expect(result.status).toBe(400);
-  });
-
-  it('writes updated config and clears cache for valid channel', async () => {
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateChannel(makeReq({ body: { channel: 'beta' } }));
-
-    expect(writeProjectReleaseChannel).toHaveBeenCalledWith('/vault', 'beta');
-    expect(clearCachedCheck).toHaveBeenCalled();
-    expect(result.status).toBeUndefined();
-  });
-
-  it('returns status from cache with exempt:false after channel change', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateChannel(makeReq({ body: { channel: 'stable' } }));
-
-    expect(result.body).toMatchObject({ exempt: false });
-  });
-
-  it('returns default status when cache is empty after channel switch', async () => {
-    vi.mocked(statusFromCache).mockReturnValue(null);
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const result = await handleUpdateChannel(makeReq({ body: { channel: 'beta' } }));
-    const body = result.body as Record<string, unknown>;
-
-    expect(body.exempt).toBe(false);
-    expect(body.update_available).toBe(false);
-    expect(body.channel).toBe('beta');
-    expect(body.channel_scope).toBe('machine');
-    expect(body.runtime_scope).toBe('machine');
-    expect(body.last_check).toBe('');
-  });
-
-  it('accepts both valid channels: stable and beta', async () => {
-    const { handleUpdateChannel } = createUpdateHandlers(makeDeps());
-
-    const stable = await handleUpdateChannel(makeReq({ body: { channel: 'stable' } }));
-    const beta = await handleUpdateChannel(makeReq({ body: { channel: 'beta' } }));
-
-    expect(stable.status).toBeUndefined();
-    expect(beta.status).toBeUndefined();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleUpdateStatus — restart_required detection
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateStatus — restart_required', () => {
-  beforeEach(() => {
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readCachedCheck).mockReturnValue(null);
-    vi.mocked(readUpdateConfig).mockReturnValue({ channel: 'stable', check_interval_hours: 6 });
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(isCacheStale).mockReturnValue(false);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    vi.mocked(getInstalledVersion).mockReset();
-    vi.mocked(getInstalledVersion).mockReturnValue(null);
-    vi.mocked(spawnRestartScript).mockReset();
-    vi.mocked(spawnRestartScript).mockReturnValue('/tmp/myco-restart-123.sh');
-  });
-
-  it('triggers auto-restart when installed version > running version', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    const scheduleShutdown = vi.fn();
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ scheduleShutdown, globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).toHaveBeenCalled();
-    expect(scheduleShutdown).toHaveBeenCalled();
-    expect(result.body).toMatchObject({ restarting: true, reason: 'version_sync' });
-  });
-
-  it('does not trigger restart when installed version matches running version', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.0.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect((result.body as Record<string, unknown>).restarting).toBeUndefined();
-  });
-
-  it('falls back to normal flow when installed version is null', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue(null);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect(result.body).toMatchObject({ exempt: false });
-  });
-
-  it('skips restart check when exempt (dev mode)', async () => {
-    vi.mocked(isUpdateExempt).mockReturnValue(true);
-    vi.mocked(getInstalledVersion).mockReturnValue('2.0.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect(result.body).toMatchObject({ exempt: true });
-  });
-
-  it('skips restart check when globalPrefix is null', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: null }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect(getInstalledVersion).not.toHaveBeenCalled();
-  });
-
-  it('does not restart when installed version is lower than running (downgrade)', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('0.9.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-  });
-
-  // Any runtime.command pin isolates the daemon from the global install.
-  // The short-circuit must skip whenever a `runtime.command` pin is set,
-  // regardless of where the pin points (dogfood `~/.local/bin/myco-dev`
-  // symlinked to a locally-built repo binary, any user-set pin). Restarting a
-  // pinned daemon would respawn the same binary and loop. The gate is purely
-  // "is a pin set?" — there is no longer a managed-runtime sub-classification.
-  it('does not trigger auto-restart when runtime.command pins a runtime', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(
-      '/vault/runtime/node_modules/.bin/myco',
-    );
-    const scheduleShutdown = vi.fn();
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ scheduleShutdown, globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect(scheduleShutdown).not.toHaveBeenCalled();
-    expect((result.body as Record<string, unknown>).restarting).toBeUndefined();
-    expect(result.body).toMatchObject({ exempt: false });
-  });
-
-  it('passes the service label into the auto-restart orchestrator when service-managed', async () => {
-    // Auto-restart short-circuit (sibling-version-sync) must route through
-    // launchd / systemd too — same thundering-herd avoidance as
-    // handleUpdateApply. Without this, the sibling daemon spawn + KeepAlive
-    // both fight for the port and the system enters the crash loop.
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    // Reset runtime.command pin pollution from preceding tests in this block.
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    const mgr = installedServiceManager(
-      'co.goondocks.myco',
-      'launchctl kickstart -k gui/501/co.goondocks.myco',
-    );
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ globalPrefix: '/usr/local', serviceManager: mgr }),
-    );
-
-    await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).toHaveBeenCalledTimes(1);
-    expect(vi.mocked(spawnRestartScript).mock.calls[0][0]).toMatchObject({
-      serviceManagedLabel: 'co.goondocks.myco',
-    });
-  });
-
-  it('does not trigger auto-restart when runtime.command pins a dogfood binary', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    // Dogfood layout: ~/.local/bin/myco-dev symlinked to the repo's built
-    // vendor binary. The daemon is pinned, so a restart would respawn the same
-    // binary and loop — the pin gate skips it.
-    vi.mocked(resolveRuntimeCommand).mockReturnValue('/Users/x/.local/bin/myco-dev');
-    const scheduleShutdown = vi.fn();
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ scheduleShutdown, globalPrefix: '/usr/local' }),
-    );
-
-    const result = await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).not.toHaveBeenCalled();
-    expect(scheduleShutdown).not.toHaveBeenCalled();
-    expect((result.body as Record<string, unknown>).restarting).toBeUndefined();
-    expect(result.body).toMatchObject({ exempt: false });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleUpdateStatus — runLocalUpdate gating against the stamp file
-//
-// The stamp file (`<vaultDir>/last-update-version`) records the version that
-// last ran `myco update` for THIS project. The version_sync restart path must
-// run `myco update --project` whenever the stamp does not already match the
-// version we are upgrading TO (the on-disk installed version) — not the
-// version we are leaving behind. The earlier predicate compared the stamp to
-// `currentVersion` (the running daemon's version), which made sibling daemons
-// silently skip the post-install sync of gitignore/symbiont/hook templates
-// after another project's daemon installed the global update first.
-// ---------------------------------------------------------------------------
-
-describe('handleUpdateStatus — runLocalUpdate stamp gating', () => {
-  let tmpVaultDir: string;
-  let tmpMycoHome: string;
-  let priorMycoHome: string | undefined;
-
-  beforeEach(() => {
-    tmpVaultDir = fs.mkdtempSync(path.join(os.tmpdir(), 'myco-update-stamp-'));
-    // Stamp now lives at `~/.myco/last-update-version` (per-machine,
-    // shared across all projects). Route MYCO_HOME at a tmp dir so the
-    // test owns the read/write path without touching the real ~/.myco.
-    tmpMycoHome = fs.mkdtempSync(path.join(os.tmpdir(), 'myco-update-stamp-home-'));
-    priorMycoHome = process.env.MYCO_HOME;
-    process.env.MYCO_HOME = tmpMycoHome;
-    vi.mocked(isUpdateExempt).mockReturnValue(false);
-    vi.mocked(readCachedCheck).mockReturnValue(null);
-    vi.mocked(readUpdateConfig).mockReturnValue({ channel: 'stable', check_interval_hours: 6 });
-    vi.mocked(readProjectReleaseChannel).mockReturnValue('stable');
-    vi.mocked(isCacheStale).mockReturnValue(false);
-    vi.mocked(statusFromCache).mockReturnValue(NO_UPDATE_STATUS);
-    vi.mocked(getInstalledVersion).mockReset();
-    vi.mocked(resolveRuntimeCommand).mockReturnValue(null);
-    vi.mocked(spawnRestartScript).mockReset();
-    vi.mocked(spawnRestartScript).mockReturnValue('/tmp/myco-restart-stamp.sh');
-  });
-
-  afterEach(() => {
-    fs.rmSync(tmpVaultDir, { recursive: true, force: true });
-    fs.rmSync(tmpMycoHome, { recursive: true, force: true });
-    if (priorMycoHome === undefined) delete process.env.MYCO_HOME;
-    else process.env.MYCO_HOME = priorMycoHome;
-  });
-
-  function writeStamp(version: string): void {
-    fs.writeFileSync(path.join(tmpMycoHome, 'last-update-version'), version, 'utf-8');
-  }
-
-  // Regression: pre-fix this returned runLocalUpdate=false because the stamp
-  // matched the running version (`currentVersion`), so sibling daemons that
-  // had previously synced their stamp to the now-outgoing version would skip
-  // `myco update --project` on every subsequent global upgrade.
-  it('runs myco update when the stamp matches the running version but not the installed version', async () => {
-    writeStamp('1.0.0');
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ vaultDir: tmpVaultDir, currentVersion: '1.0.0', globalPrefix: '/usr/local' }),
-    );
-
-    await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).toHaveBeenCalledTimes(1);
-    expect(vi.mocked(spawnRestartScript).mock.calls[0][0]).toMatchObject({
-      runLocalUpdate: true,
-      fromVersion: '1.0.0',
-      toVersion: '1.1.0',
-    });
-  });
-
-  it('skips myco update when the stamp already matches the installed (target) version', async () => {
-    writeStamp('1.1.0');
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ vaultDir: tmpVaultDir, currentVersion: '1.0.0', globalPrefix: '/usr/local' }),
-    );
-
-    await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).toHaveBeenCalledTimes(1);
-    expect(vi.mocked(spawnRestartScript).mock.calls[0][0]).toMatchObject({
-      runLocalUpdate: false,
-    });
-  });
-
-  it('runs myco update when the stamp file is missing (fresh install)', async () => {
-    vi.mocked(getInstalledVersion).mockReturnValue('1.1.0');
-    const { handleUpdateStatus } = createUpdateHandlers(
-      makeDeps({ vaultDir: tmpVaultDir, currentVersion: '1.0.0', globalPrefix: '/usr/local' }),
-    );
-
-    await handleUpdateStatus(makeReq());
-
-    expect(spawnRestartScript).toHaveBeenCalledTimes(1);
-    expect(vi.mocked(spawnRestartScript).mock.calls[0][0]).toMatchObject({
-      runLocalUpdate: true,
-    });
-  });
-});
